further bmp tests

+++ src/graphics/Square.java
@@ -5,7 +5,7 @@ import org.lwjgl.util.vector.Vector2f;
 public class Square extends BasicRenderObject {
 	private Rectangle rectangle;
 
-	private int length;
+	private float length;
 
 	/**
 	 * Default Square length is 1, just like a pixel
@@ -14,12 +14,12 @@ public class Square extends BasicRenderObject {
 		this(1);
 
 	}
-	public Square(int length) {
+	public Square(float length) {
 		this.rectangle = new Rectangle(length, length);
 		this.length = length;
 
 	}
-	public Square(int length, float r, float g, float b) {
+	public Square(float length, float r, float g, float b) {
 		this(length);
 		this.setColor(r, g, b);
 	}
